+++ CHANGELOG.md
@@ -1,10 +1,12 @@
-## 3.0.0-dev
+## 3.0.0
 
 * **Breaking change:** Remove `ListSyntax.removeLeadingEmptyLine`,
   `ListSyntax.removeTrailingEmptyLines`, `TableSyntax.parseAlignments`,
   `TableSyntax.parseRow`.
 * Allow intra-word strikethrough in GFM
   ([#300](https://github.com/dart-lang/markdown/issues/300)).
+* **Breaking change:** Change `BlockSyntax.canEndBlock` from a getter to a
+  method accepting a BlockParser.
 
 ## 2.1.8